implement logging but deactivate for now
Logging is hard to do in parallel because with logging.basicConfig
multiple engines would write to the same file. Maybe the way forward
is to use the grepLogger (or similar) so that every instance uses the
same logging but in a controlled way. (Like when several modules use
the same logging instance).