Some notebooks supporting the final data products
ClusteringManager is where I produce the clustered data outputs
Ground projection has 2 parts, first to download and process the HiRISE
data, the second part for the actual xy2latlon, which has become quite
more complicated, now that I need to do it for several points per
marking cluster.
Convert existing catalog I used when I wanted to convert the format of
the existing cluster output to help for ground projection. Kinda
obsolete now but might be helpful for similar things in the future.