Removing clustering from plotting of pipeline
The point is, I don’t want the plotter to be able to execute the
pipeline,
so even so it’s more code in the notebook, I rather have it separated
for
better control.
I now also check if things exist before I plot, so that it doesn’t
choke in that case.