implement a more rigorous scope management
Blotch and Fan now require a scope to be set and use the correct columns
for base or center coordinates appropiately. This solves most of GH53
but there is a remaining overlap issue between tiles when clustering
in HiRise scope. I commit this so that we can evaluate how much overlap
issues exist.